+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 All notable changes to this project will be documented in this file. This project does not adhere to [Semantic Versioning](http://semver.org/)
 ## 2.2.3 (07/07/2016)
 - QA-246: Dupe attempt test for SCRA
+- QA-244: Dupe attempt test for Military Document
+- QA-248: Dupe attempt test for EMT State Lookup
 
 ## 2.2.2 (07/05/2016)
 - QA-294: Add spec for marketplace reviews